result filter adjusted

+++ result/result_filter.py
@@ -1,6 +1,7 @@
 import os
 import json
 import shutil
+distribution = {}
 def check_and_delete_folders():
     # Get current directory
     current_dir = os.getcwd()
@@ -15,7 +16,7 @@ def check_and_delete_folders():
         # Check if it's a directory
         if os.path.isdir(folder_path):
             score_file = os.path.join(folder_path, 'score.json')
-            
+            DM_file = os.path.join(folder_path, 'DM_query.json')
             # Check if score file exists
             if os.path.exists(score_file):
                 try:
@@ -27,10 +28,21 @@ def check_and_delete_folders():
                         # Delete the folder
                         shutil.rmtree(folder_path)
                         deleted_folders.append(folder)
-                        
+                    # else:
+                    #     config_info = folder_path.split('_')
+                    #     if config_info[0] not in distribution:
+                    #         distribution[config_info[0]] = 1
+                    #     else:
+                    #         distribution[config_info[0]] += 1
+                    #     if config_info[0] == "interact":
+                    #         if config_info[1] not in distribution:
+                    #             distribution[config_info[1]] = 1
+                    #         else:
+                    #             distribution[config_info[1]] += 1
+
                 except (json.JSONDecodeError, FileNotFoundError) as e:
                     print(f"Error reading {score_file}: {e}")
-            else:
+            elif os.path.exists(DM_file):
                 shutil.rmtree(folder_path)
                 deleted_folders.append(folder)
 
@@ -45,3 +57,5 @@ def check_and_delete_folders():
 # Run the function
 if __name__ == "__main__":
     check_and_delete_folders()
+    # for key in distribution.keys():
+    #     print(key, distribution[key])
